api/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
supabase
replicate
pillow
requests
pyyaml
deep-translator
jsonschema
pydantic
python-slugify